Ignore some updates from virt driver
During update_available_resource periodic task, compute node object
in the database is first updated by resource tracker's _update method
with the values reported from the system by virt driver's
get_available_resource; then it is updated with instance usages by the
same method. In both cases, _update method saves the currently written
usages to old_resources dictionary to check if new update brings any
changes, and skip it if it does not. The problem is, system usages
are very unlikely to be the same as instance usages, so the resources
update both in database and in placement are always going to be updated.
It can be avoided by the proposed fix.

For "vcpus_used", "memory_mb_used", "local_gb_used", nova compute
should calculate the value according to the booted instances, so we
can skip updating these vaules from virt driver.

Closes-Bug: #1733034
Change-Id: I5471c30f233fe4acaf1f9b545f91e09af1e60137