Add method to generate device names universally
When attaching block devices, we find the next device name to use by
looking at the used device letters in the block device mappings and
selecting the next available letter, as long as the letter is within
'a' - 'z' or 'aa' - 'zz'. This allows us to represent 702 devices (and
would interestingly hit an IndexError if all 'a' - 'z' and 'aa' - 'zz'
letters were in use already.

This adds a method for generating device names universally, not limited
to any particular maximum, based on a device index. For example:

  vda(index=0), sdb(index=1), vdaa(index=26), sdabc(index=730), ...

This method and its helper method will be used in a subsequent patch.

Part of blueprint conf-max-attach-volumes

Change-Id: I2699f600a98222a4dc58903ec97ca70e57701c93