Fix attaching port with different port security
If the port_security_enabled setting between the port being attached and
it's network differ, then the setting from the port should be honoured.
The setting on the network is only a default value for ports created in
this network without an explicit port security setting. The decisive
setting is that of the port.

Closes-Bug: 2056544
Change-Id: Ie2608918bfcfe664634c947f176c0a6422389aa4