Always reuse the driver element for live-migration
Changes to the config, or to Nova behavior, may cause the driver elemnt
in the libvirt xml config to change. VMs created before the config or
behavior change, can then no longer be live-migrated.

By always copying the driver element, compatibility between source and
target XML is increased and live-migration issues are avoided.