Resize ephemeral disk on instance resize
Block device mappings for ephemeral and swap disks are not resized during
instance resize operations. This causes virt drivers to not resize them.

This patch resizes bdms for ephemeral and swap disks during
finish_resize if:
- All space allowed by the old flavor is used.
- There is a single ephemeral disk or no ephemeral disk.
- Swap is reduced in all cases if the new flavor allows less swap than
  currently allocated.
For all other cases, the disk configuration is not touched as it's
unclear how to resize.

Note: Resizing to a flavor with less ephemeral disk space is not
allowed and already prohibited in the API layer.

TODO from old change:
Virt driver related changes:
- libvirt: a new imagebackend method _get_original_template_name is
added, which aim is to get template name of an existing disk, which
differs from specified one from driver's _create_image method in the
case of resizing. Really, if we resize an ephemeral disk from 10 GB to
20 GB, the bdm object is already has 20 GB size, and _create_image
specifies template name for 20 GB disk, but existing disk is created
from 10 GB template and must continue to work with it. There is no
reason to charge driver with such details, so that the new virtual
method of imagebackend is the best choice.

Resize ephemerals down details by drivers:
- hyperv prevents resizing down in finish_migration method (except
resizing to zero, in which case it removes ephemeral disk). This method
raises an exception, and the instance is goes to error state.
Unfortunately the same occures with revert of resize up operation when
ephemeral bdm size is decreasing. I.e. it looks like this patch brings a
problem there;
- vmwareapi does not honor ephemeral bdms at all, but works with flavor
ephemeral size directly. Moreover it recreates ephemeral disks on any
resize operation. So that it probably does need to do something
additionally with this patch;
- ironic does not seem to has resizing supported at all.
Conclusions: if vmwareapi behavior (recreating of ephemerals) is legal,
fix hyperv only; if not - an early performed common check preventing
resizing down whould be suitable.

TODOes:
- fix handling of resized bdms in hyperv;
- check against lvm and ploop image backends;
- fix a bug with disk.info inconsistency for flat backend after resizing
(disks becomes qcow2 while disk.info still has 'raw' for them);

Change-Id: Ie42a0b2f0019202c6cb1a1533a838a8549c000c8
Related-Bug: 1558880
Related-Bug: 1605720
(cherry picked from commit dcabe3588afeade262a4ec14d1cc7067d1f0ef0e)